timer added, fps bug solved
