+++ src/com/codecool/snake/Snake.java
@@ -1,0 +1,19 @@
+package com.codecool.snake;
+
+import javafx.application.Application;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+public class Snake extends Application {
+
+    @Override
+    public void start(Stage primaryStage) {
+        Game game = new Game();
+
+        primaryStage.setTitle("Snake Game");
+        primaryStage.setScene(new Scene(game, Globals.WINDOW_WIDTH, Globals.WINDOW_HEIGHT));
+        primaryStage.show();
+        game.start();
+    }
+
+}
